// src/shared/types/inputLogin.ts
import { JSX } from 'react';

export type InputProps = {
	errorMessage?: string; //Текст сообщения об ошибке, который отображается под полем, если поле содержит ошибку
	warnMessage?: string; // Текст предупреждающего сообщения, который отображается под полем, если есть предупреждение
	error?: boolean; // Флаг, указывающий, что поле содержит ошибку. Используется для стилизации и управления отображением errorMessage.
	warn?: boolean; // Флаг, указывающий, что поле содержит предупреждение. Используется для стилизации и управления отображением warnMessage
	label?: string; // Текст метки (лейбла) поля ввода, который отображается над инпутом.
	indicator?: number; // Числовой индикатор состояния поля. Используется для отображения визуального индикатора рядом с полем 0-дефолт, 1-ошибка, 2-предупреждение, 3-ОК
	done?: boolean; // Флаг, показывающий, что поле успешно заполнено и валидно.
} & JSX.IntrinsicElements['input'];